docs, removed redundant chart interface
