+++ plotlyjs-api/src/main/java/net/java/html/plotlyjs/HistogramMarker.java
@@ -40,13 +40,11 @@ package net.java.html.plotlyjs;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.Map;
 
-/**
- *
- * @author daykin
- */
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class HistogramMarker extends Marker{
 
+    
     public static class Builder {
 
         private Boolean autocolorscale;
@@ -66,70 +64,85 @@ public class HistogramMarker extends Marker{
         private Builder() {
         }
 
+        
         public Builder autocolorscale(final Boolean value) {
             this.autocolorscale = value;
             return this;
         }
 
+        
         public Builder cmax(final Number value) {
             this.cmax = value;
             return this;
         }
 
+        
         public Builder cmin(final Number value) {
             this.cmin = value;
             return this;
         }
 
+        
         public Builder colorscale(final Map<Number,String> value) {
             this.colorscale = value;
             return this;
         }
 
+        
         public Builder color(final String value) {
             this.color = value;
             return this;
         }
 
+        
         public Builder reversescale(final Boolean value) {
             this.reversescale = value;
             return this;
         }
 
+        
         public Builder cauto(final Boolean value) {
             this.cauto = value;
             return this;
         }
 
+        
         public Builder colorbar(final ColorBar value) {
             this.colorbar = value;
             return this;
         }
 
+        
         public Builder line(final Line value) {
             this.line = value;
             return this;
         }
 
+        
         public Builder colorsrc(final String value) {
             this.colorsrc = value;
             return this;
         }
 
+        
         public Builder showscale(final Boolean value) {
             this.showscale = value;
             return this;
         }
         
+        
         public Builder opacity(final Number value) {
             this.opacity = value;
             return this;
         }
+
+        
         public HistogramMarker build() {
             return new net.java.html.plotlyjs.HistogramMarker(this);
         }
     }
 
+    
     public static HistogramMarker.Builder builder() {
         return new HistogramMarker.Builder();
     }
