+++ plotlyjs-api/src/main/java/net/java/html/plotlyjs/HoverEvent.java
@@ -32,16 +32,15 @@ package net.java.html.plotlyjs;
  */
 
 
-/**
- *
- * @author daykin
- */
+
 public class HoverEvent extends ChartEvent{
     
-    public HoverEvent(Plotly source, Object shift, Object ctrl, Object info) {
+    
+    public HoverEvent(Plotly<?> source, Object shift, Object ctrl, Object info) {
         super(source, shift, ctrl, info);
         //other click related stuff
     }
+    
     
     public String getLabel(){
         return ((Plotly)source).getId();
