+++ plotlyjs-api/src/main/java/net/java/html/plotlyjs/Scatter.java
@@ -1,5 +1,4 @@
 package net.java.html.plotlyjs;
-
 /*
  * #%L
  * This software is Copyright by the Board of Trustees of Michigan State University.
@@ -30,10 +29,12 @@ package net.java.html.plotlyjs;
  * THE SOFTWARE.
  * #L%
  */
-
 import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.List;
-
+/**
+ * @author daykin
+ * @param <TRACE> 
+ */
 @SuppressWarnings("unused")
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class Scatter<TRACE extends Trace> extends Charts{
@@ -71,7 +72,9 @@ public class Scatter<TRACE extends Trace> extends Charts{
     private final String name;
     private final Boolean connectgaps;
     private final List<Number> r;
-    
+    /**
+     * @param <TRACE_2>
+     */
     public static class Builder<TRACE_2 extends Trace> {
         private List<?> x;
         private List<?> y;
@@ -107,13 +110,10 @@ public class Scatter<TRACE extends Trace> extends Charts{
         private String name;
         private Boolean connectgaps;
         private List<Number> r;
-        
         private Builder(){}
-        
         private static <TRACE_3 extends Trace> Builder<TRACE_3> start(){
             return new Builder<>();
         }
-        
         public Builder<TRACE_2> trace(final TRACE_2 tr){
             if (tr instanceof PolarTrace){
                this.r = ((PolarTrace)tr).r;
@@ -125,176 +125,145 @@ public class Scatter<TRACE extends Trace> extends Charts{
             }
             return this;
         }
-           
         public Builder<TRACE_2> mode(final String value) {
             this.mode = value;
             return this;
         }
-
         public Builder<TRACE_2> textposition(final String value) {
             this.textposition = value;
             return this;
         }
-
         public Builder<TRACE_2> stream(final Stream value) {
             this.stream = value;
             return this;
         }
-
         public Builder<TRACE_2> text(final String value) {
             this.text = value;
             return this;
         }
-
         public Builder<TRACE_2> hoverinfo(final String value) {
             this.hoverinfo = value;
             return this;
         }
-
         public Builder<TRACE_2> xsrc(final String value) {
             this.xsrc = value;
             return this;
         }
-
         public Builder<TRACE_2> visible(final Object value) {
             this.visible = value;
             return this;
         }
-
         public Builder<TRACE_2> marker(ScatterMarker value) {
             this.marker = value;
             return this;
         }
-
         public Builder<TRACE_2> y0(final Number value) {
             this.y0 = value;
             return this;
         }
-
         public Builder<TRACE_2> tsrc(final String value) {
             this.tsrc = value;
             return this;
         }
-
         public Builder<TRACE_2> line(final Line value) {
             this.line = value;
             return this;
         }
-
         public Builder<TRACE_2> fill(final String value) {
             this.fill = value;
             return this;
         }
-
         public Builder<TRACE_2> showlegend(final Boolean value) {
             this.showlegend = value;
             return this;
         }
-
         public Builder<TRACE_2> error_x(final ErrorBar value) {
             this.error_x = value;
             return this;
         }
-
         public Builder<TRACE_2> error_y(final ErrorBar value) {
             this.error_y = value;
             return this;
         }
-
         public Builder<TRACE_2> textsrc(final String value) {
             this.textsrc = value;
             return this;
         }
-
         public Builder<TRACE_2> rsrc(final String value) {
             this.rsrc = value;
             return this;
         }
-
         public Builder<TRACE_2> xaxis(final Axis value) {
             this.xaxis = value;
             return this;
         }
-
         public Builder<TRACE_2> yaxis(final Axis value) {
             this.yaxis = value;
             return this;
         }
-
         public Builder<TRACE_2> ysrc(final String value) {
             this.ysrc = value;
             return this;
         }
-
         protected Builder<TRACE_2> t(final List<Number> value) {
             this.t = value;
             return this;
         }
-
         public Builder<TRACE_2> opacity(final Number value) {
             this.opacity = value;
             return this;
         }
-
         public Builder<TRACE_2> textfont(final Font value) {
             this.textfont = value;
             return this;
         }
-
         public Builder<TRACE_2> legendgroup(final String value) {
             this.legendgroup = value;
             return this;
         }
-
         public Builder<TRACE_2> textpositionsrc(final String value) {
             this.textpositionsrc = value;
             return this;
         }
-
         public Builder<TRACE_2> fillcolor(final String value) {
             this.fillcolor = value;
             return this;
         }
-
         public Builder<TRACE_2> dx(final Number value) {
             this.dx = value;
             return this;
         }
-
         public Builder<TRACE_2> dy(final Number value) {
             this.dy = value;
             return this;
         }
-
         public Builder<TRACE_2> x0(final Number value) {
             this.x0 = value;
             return this;
         }
-
         public Builder<TRACE_2> name(final String value) {
             this.name = value;
             return this;
         }
-
         public Builder<TRACE_2> connectgaps(final Boolean value) {
             this.connectgaps = value;
             return this;
         }
-
         protected Builder<TRACE_2> r(final List<Number> value) {
             this.r = value;
             return this;
         }
-        
         public Scatter<TRACE_2> build() {
             return new Scatter<>(this);
         }                
     }
-    
+    /**
+     * @param <TRACE>
+     * @return a builder for a scatter plot.
+     */
     public static <TRACE extends Trace> Builder<TRACE> builder(){
         return Scatter.Builder.start();
     }
-    
     private Scatter(Builder<TRACE> builder){
         this.x = builder.x;
         this.y = builder.y;
@@ -331,6 +300,4 @@ public class Scatter<TRACE extends Trace> extends Charts{
         this.connectgaps = builder.connectgaps;
         this.r = builder.r;    
     }
-
-
 }
