--- plotlyjs-api/src/main/java/net/java/html/plotlyjs/package.html
@@ -1,104 +1,0 @@
-<!--
-  #%L
-  Charts API for HTML/Java - a library from the "DukeScript" project.
-  %%
-  Copyright (C) 2015 Dukehoff GmbH
-  %%
-  Permission is hereby granted, free of charge, to any person obtaining a copy
-  of this software and associated documentation files (the "Software"), to deal
-  in the Software without restriction, including without limitation the rights
-  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-  copies of the Software, and to permit persons to whom the Software is
-  furnished to do so, subject to the following conditions:
-  
-  The above copyright notice and this permission notice shall be included in
-  all copies or substantial portions of the Software.
-  
-  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-  THE SOFTWARE.
-  #L%
-  -->
-<!DOCTYPE html>
-<html>
-    <head>
-        <title></title>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-        <style>
-            body {
-                -webkit-touch-callout: none;
-                -webkit-user-select: none;
-                -khtml-user-select: none;
-                -moz-user-select: none;
-                -ms-user-select: none;
-                user-select: none;
-            }
-        </style>
-    </head>
-    <body>
-        <p>
-            Create various types of {@link net.java.html.charts.Chart charts}.
-            Display them
-            using <a href="http://bits.netbeans.org/html+java/">HTML/Java</a>
-            API and other presenters provided by the
-            <a href="http://dukescript.com">DukeScript</a> project.
-        </p>
-    <center>
-        <button data-bind="click: bar">Bar</button>
-        <button data-bind="click: pie">Pie</button>
-        <button data-bind="click: line">Line</button>
-        <button data-bind="click: doughnut">Doughnut</button>
-        <button data-bind="click: radar">Radar</button>
-        <button data-bind="click: polar">Polar</button>
-        <canvas id="chart"></canvas>
-        <p>
-            Click the chart to increase the values. Hold <em>Shift</em>
-            to decrease, hold <em>Ctrl</em> to modify the value by ten.
-        </p>
-        <textarea rows="15" data-bind="value: code" style="width: 100%">
-        </textarea>
-    </center>
-    <p>
-        To use this library from your
-        <a target="_blank" href="http://dukescript.com">DukeScript</a> application follow
-        the <a target="_blank" href="https://dukescript.com/getting_started.html">getting started tutorial</a>
-        and then add dependency on the <b>Charts API</b> to your
-        <code>client/pom.xml</code>:
-    </p>
-    <pre>
-&lt;dependency&gt;
-  &lt;groupId&gt;com.dukescript.api&lt;/groupId&gt;
-  &lt;artifactId&gt;charts&lt;/artifactId&gt;
-  &lt;version&gt;0.8&lt;/version&gt; &lt-- or any newer --&gt;
-  &lt;type&gt;jar&lt;/type&gt;
-&lt;/dependency&gt;
-    </pre>
-    <p>
-        If you are also using <code>client-web/pom.xml</code> consider to include
-        pre-built <a target="_blank" href="http://bck2brwsr.apidesign.org">Bck2Brwsr Library</a>
-        as
-    </p>
-    <pre>
-&lt;dependency&gt;
-  &lt;groupId&gt;com.dukescript.api&lt;/groupId&gt;
-  &lt;artifactId&gt;charts&lt;/artifactId&gt;
-  &lt;version&gt;0.8&lt;/version&gt; &lt-- the same as above --&gt;
-  &lt;classifier&gt;bck2brwsr&lt;/classifier&gt;
-&lt;/dependency&gt;
-    </pre>
-    <p>
-        Including this pre-built library will speed your compilation time up.
-    </p>
-
-<!-- initialize TeaVM -->
-<script type="text/javascript" src="doc-files/teavm.js"></script>
-<script>
-    main();
-</script>
-<!-- end of initialization -->
-    </body>
-</html>
